requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gradio>=4.0.0
openai>=1.30.0
python-dotenv>=1.0.0
langchain>=0.1.0
transformers>=4.36.0
torch>=2.1.0
numpy>=1.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
watchfiles>=0.21.0
chromadb>=0.4.24
sentence-transformers>=2.2.2
tqdm>=4.66.0
scikit-learn>=1.3.0
websocket-client>=1.6.0
pandas>=2.0.0
pyarrow>=14.0.0